requirements.txt
# requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
